change return type

+++ 0x13-more_singly_linked_lists/101-print_listint_safe.c
@@ -8,7 +8,7 @@
  */
 size_t print_listint_safe(const listint_t *head)
 {
-	int count = 0;
+	size_t count = 0;
 
 	while (head != NULL)
 	{
